GetQuantityTrendAnalysisItem decimal fix

+++ src/DashboardTsy.Application/TargetReport/Responses/GetQuantityTrendAnalysisItem.cs
@@ -7,5 +7,5 @@ public class GetQuantityTrendAnalysisItem
 {
     public string ProductName { get; set; } = string.Empty;   // SP: URUN
     public DateTime ReportDate { get; set; }                   // SP: TARIH_DETAY
-    public long Count { get; set; }                            // SP: ADET
+    public decimal Count { get; set; }                         // SP: ADET
 }
